BUG/MEDIUM: connections: Add the CO_FL_CONNECTED flag if a send succeeded.
If a send succeeded, add the CO_FL_CONNECTED flag, the send may have been
called by the upper layers before we even realized we were connected, and we
may even read the response before we get the information, and si_cs_recv()
has to know if we were connected or not.

This should be backported to 1.9.

(cherry picked from commit f24502ba461d2dff1268aa73ce70a3400909d8d7)
Signed-off-by: Willy Tarreau <w@1wt.eu>