BUG/MINOR: mux-h1: Apply the reserve on the channel's buffer only
There is no reason to use the reserve to limit the amount of data of the input
buffer that we can parse at a time. The only important thing is to keep the
reserve free in the channel's buffer.

This patch should be backported to 1.9.

(cherry picked from commit afe57846bfb2edfb0db53f942b9500d090377c54)
Signed-off-by: Willy Tarreau <w@1wt.eu>