BUG/MINOR: stream: take care of synchronous errors when trying to send
We currently detect a number of situations where we have to immediately
deal with a state change, but we failed to consider the case of the
synchronous error reported on the stream-interface. We definitely do not
want to have to wait for a timeout to handle this one, especially at the
beginning of the connection when it can lead to an immediate retry.

This should be backported to 1.9.

(cherry picked from commit 64f6945fec20a02d90c647c7426a989b903f6f5d)
Signed-off-by: Willy Tarreau <w@1wt.eu>