MINOR: stream: don't wait before retrying after a failed connection reuse
When a connection reuse fails, we must not wait before retrying, as most
likely the issue is related to the reused connection and not to the server
itself.

This should be backported to 1.9, though it depends on previous patches
dealing with SI_ST_CON for connection reuse.

(cherry picked from commit f7a259d46f84ade7222debbce937d5d94bc568f4)
Signed-off-by: Willy Tarreau <w@1wt.eu>