MEDIUM: h2: always parse and deduplicate the content-length header
The header used to be parsed only in HTX but not in legacy. And even in
HTX mode, the value was dropped. Let's always parse it and report the
parsed value back so that we'll be able to store it in the streams.

(cherry picked from commit 4790f7c90743ce52021373c5567d78c32c42d972)
Signed-off-by: Willy Tarreau <w@1wt.eu>