BUG/MINOR: mux-h2: always compare content-length to the sum of DATA frames
This is mandated by RFC7541#8.1.2.6. Till now we didn't have a copy of
the content-length header field. But now that it's already parsed, it's
easy to add the check.

The reg-test was updated to match the new behaviour as the previous one
expected unadvertised data to be silently discarded.

This should be backported to 1.9 along with previous patch (MEDIUM: h2:
always parse and deduplicate the content-length header) after it has got
a bit more exposure.

(cherry picked from commit 1915ca273832ba542d72eb0645dd7ccb6d5b945f)
[wt: removed the accidental debugging printf]
Signed-off-by: Willy Tarreau <w@1wt.eu>