BUG/MINOR: listener: always fill the source address for accepted socketpairs
The source address was not set but passed down the chain to the upper
layer's accept() calls. Let's initialize it like other UNIX sockets in
this case. At the moment it should not have any impact since socketpairs
are only usable for the master CLI.

This should be backported to 1.9.

(cherry picked from commit 888d5678f7829e0685d0f692cdb159e655ac0e25)
Signed-off-by: Willy Tarreau <w@1wt.eu>