BUG/MEDIUM: servers: Don't add an incomplete conn to the server idle list.
If we failed to add the connection to the session, only try to add it back
to the server idle list if it has a mux, otherwise the connection is
incomplete and unusable.

This should be backported to 1.9.

(cherry picked from commit 26da323cb9f72c34062a3f023950da100f4fef55)
Signed-off-by: Willy Tarreau <w@1wt.eu>