MINOR: h2: add a generic frame checker
The new function h2_frame_check() checks the protocol limits for the
received frame (length, ID, direction) and returns a verdict made of
a connection error code. The purpose is to be able to validate any
frame regardless of the state and the ability to call the frame handler,
and to emit a GOAWAY early in this case.

(cherry picked from commit 9c84d8299a733bf9bb979cd147e9c566f9e1ae67)
Signed-off-by: Willy Tarreau <w@1wt.eu>