DOC: ssl: Stop documenting ciphers example to use
Since TLS ciphers are not well understand, it is very common pratice to
copy and paste parameters from documentation and use them as-is. Since RC4
should not be used anymore, it is wiser to link users to up to date
documnetation from Mozilla to avoid unsafe configuration in the wild.

Clarify the location of man pages for OpenSSL when missing.

(cherry picked from commit 4f03ab06a90df8e88ba2e347f52465b31392acc4)
Signed-off-by: Willy Tarreau <w@1wt.eu>