BUG/MAJOR: spoe: verify that backends used by SPOE cover all their callers' processes
When a filter is installed on a proxy and references spoe, we must be
absolutely certain that the whole chain is valid on a given process
when running in multi-process mode. The problem here is that if a proxy
1 runs on process 1, referencing an SPOE agent itself based on a backend
running on process 2, this last one will be completely deinited on
process 1, and will thus cause random crashes when it gets messages
from this proess.

This patch makes sure that the whole chain is valid on all of the caller's
processes.

This fix must be backported to all spoe-enabled maintained versions. It
may potentially disrupt configurations which already randomly crash.
There hardly is any intermediary solution though, such configurations
need to be fixed.

(cherry picked from commit 2bdcfde4260ac9115b8a0b7aa916975799273ea9)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit b54a86df6f12ac82614921a8c0d4078d781b1ea7)
Signed-off-by: Willy Tarreau <w@1wt.eu>