BUG/MAJOR: spoe: Don't try to get agent config during SPOP healthcheck
During SPOP healthchecks, a dummy appctx is used to create the HAPROXY-HELLO
frame and then to parse the AGENT-HELLO frame. No agent are attached to it. So
it is important to not rely on an agent during these stages. When HAPROXY-HELLO
frame is created, there is no problem, all accesses to an agent are
guarded. This is not true during the parsing of the AGENT-HELLO frame. Thus, it
is possible to crash HAProxy with a SPOA declaring the async or the pipelining
capability during a healthcheck.

This patch must be backported to 1.9 and 1.8.

(cherry picked from commit 11389018bc92bf7b94533e682af5cb4bbf0e43d9)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 644283b87d183ad3a22c3f3e0145bde8387b4e6e)
Signed-off-by: Willy Tarreau <w@1wt.eu>