BUG/MINOR: config: Reinforce validity check when a process number is parsed
Now, in the function parse_process_number(), when a process number or a set of
processes is parsed, an error is triggered if an invalid character is found. It
means following syntaxes are not forbidden and will emit an alert during the
HAProxy startup:

  1a
  1/2
  1-2-3

This bug was reported on Github. See issue #36.

This patch may be backported to 1.9 and 1.8.

(cherry picked from commit 18cca781f5384f060704ad80018d80bdd4e01e76)
[wt: adjusted context, s/max/LONGBITS]
Signed-off-by: Willy Tarreau <w@1wt.eu>

(cherry picked from commit adadecb56f68e173b19b0aa28225812e463d6930)
Signed-off-by: Willy Tarreau <w@1wt.eu>