BUILD/MINOR: peers: remove an impossible null test in intencode()
intencode() tests for the nullity of the target pointer passed in
argument, but the code calling intencode() never does so and happily
dereferences it. gcc at -O3 detects this as a potential null deref.
Let's remove this incorrect and misleading test. If this pointer was
null, the code would already crash in the calling functions.

This must be backported to stable versions.

(cherry picked from commit 9bdd7bc63d94501e649bc0f5c0c3d4a4288004cc)
Signed-off-by: Willy Tarreau <w@1wt.eu>